finished 64bit port

+++ arch/x86_64/include/ArchThreads.h
@@ -127,7 +127,7 @@ public:
  * @param new_value to set variable lock to
  * @returns old_value of variable lock
  */
-  static uint32 testSetLock(uint32 &lock, uint32 new_value);
+  static size_t testSetLock(size_t &lock, size_t new_value);
 
 /**
  * atomically increments or decrements value by increment
